+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- fix the lowercasing of password bug.
+
 ## [1.1.3] - 2025-05-21
 
 - Make dirs a global dependency.